requirements.txt
torch
torchvision
facenet-pytorch
pillow
matplotlib
numpy